refactor: move domain objects to model module
