+++ src/db/sqlx/record.rs
@@ -1,8 +1,7 @@
 use chrono::{Datelike, NaiveDateTime};
 use sqlx::prelude::FromRow;
 
-use crate::{domain::income::Income, time::Quarter};
-
+use crate::{domain::Income, time::Quarter};
 #[derive(Debug, Clone, FromRow)]
 pub struct IncomeRecord {
     pub date: NaiveDateTime,
