+++ src/domain/model/mod.rs
@@ -1,0 +1,4 @@
+//! Definitions of domain entities and values.
+pub mod income;
+pub mod payment;
+pub mod tax_payment;
